feat: Add team members' images and update Our Team page
This commit adds the images of team members to the Our Team page. It also updates the page to include the North Studio Managing Directors and Core Members as demo data. Additionally, the CEO is added to the North Studio Family Members section. The TeamCard component is created and added to the page.

Fixes : default data